feat: Start adding unit tests. (#29)
Add tests for main.go with a first unit test for the function
"getBackupActionset". The test instantiates a fake dynamic
client, registers some ActionSets into it and then passes
it to the function. The test then verifies that the expected
backups are returned from the function.

The commit also includes a Github workflow to run the tests
upon PR creation and merge to main.